pyproject.toml
[project]
name = "flowboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "django-background-tasks==1.2.8",
    "django-compat==1.0.15",
    "httpx>=0.28.1",
    "pillow>=12.1.0",
    "psycopg[binary]>=3.2.3",
    "python-decouple>=3.8",
    "python-dotenv>=1.2.1",
]
